+++ crates/aspen-raft/src/verified/conversion.rs
@@ -264,7 +264,7 @@ pub fn encode_batch_op(op: &CompactBatchOp) -> (bool, String, String) {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct BatchSizeError {
     /// Actual batch size.
-    pub size: usize,
+    pub size: u32,
     /// Maximum allowed size.
     pub max: u32,
 }
@@ -290,8 +290,8 @@ pub struct BatchSizeError {
 /// assert!(validate_batch_size(150, 100).is_err());
 /// ```
 #[inline]
-pub fn validate_batch_size(batch_size: usize, max_allowed: u32) -> Result<(), BatchSizeError> {
-    if batch_size > max_allowed as usize {
+pub fn validate_batch_size(batch_size: u32, max_allowed: u32) -> Result<(), BatchSizeError> {
+    if batch_size > max_allowed {
         Err(BatchSizeError {
             size: batch_size,
             max: max_allowed,
